refactor(companion): extract BLETransport, fold WiFi/Lichess managers into BoardConnection
